backend/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchaudio
coqui-tts
f5-tts-mlx
pydub
numpy<2.0.0
huggingface_hub
python-dotenv
scipy
librosa
openai-whisper
whisperx
# transformers 4.44.2 is the sweet spot:
#   - whisperx requires >=4.39.0
#   - coqui-tts (XTTS) requires BeamSearchScorer, removed in 4.45+
transformers==4.44.2
torchcodec